Ignore these changes

+++ flowscale/inc/mainwindow.h
@@ -37,7 +37,7 @@ public:
 
     std::ofstream fout;
 
-    int *_beltRoundCounter;
+    unsigned long *_beltRoundCounter;
     int _weightValueFromScale;
     long _counter;
     int _step;
@@ -72,7 +72,7 @@ private slots:
     // /////////////////////////
 
 public slots:
-    void conveyorBeltCounter(int);
+    void conveyorBeltCounter(unsigned long);
     void recordWeight(int);
     void displayReceivedWeight(int);
     void displayInputValue(unsigned long);
