+++ flowscale/src/mainwindow.cpp
@@ -30,7 +30,7 @@ MainWindow::MainWindow(QWidget *parent) :
     connect(scale, SIGNAL(receivedWeight(int)),         this, SLOT(recordWeight(int)));
     connect(dio,   SIGNAL(inputValue(unsigned long)),   this, SLOT(displayInputValue(unsigned long)));
 
-    connect(dio,   SIGNAL(conveyorSignal(unsigned long)),  this, SLOT(conveyorBeltCounter(int)));
+    connect(dio,   SIGNAL(conveyorSignal(unsigned long)),  this, SLOT(conveyorBeltCounter(unsigned long)));
     //connect(dio,   SIGNAL(tachoSignal(unsigned long)),  this, SLOT(displayTachoCount(unsigned long)));
 
     // /////////////////////////
@@ -49,7 +49,7 @@ MainWindow::~MainWindow()
     delete _beltRoundCounter;
 }
 
-void MainWindow::conveyorBeltCounter(int beltRoundCounter)
+void MainWindow::conveyorBeltCounter(unsigned long beltRoundCounter)
 {
     // do nothing, just receive value to pointer temporarily
     this->_beltRoundCounter = &beltRoundCounter;
